End-of-day commit.

+++ core/Response.h
@@ -4,7 +4,7 @@
 
 namespace Gadgetron::Core {
     struct Response {
-        Response(uint16_t correlation_id, const std::string &string);
+        Response(uint64_t correlation_id, const std::string &string);
 
         uint64_t correlation_id;
         std::string response;
